refactor(ports): relocate metrics format funcs to internal/pkg/metricsfmt (issue #1343 Phase 1)
