+++ internal/agent/session/session_manager.go
@@ -17,6 +17,7 @@ import (
 	"github.com/gosharplite/tell-me-go/internal/domain/ports"
 	domain_security "github.com/gosharplite/tell-me-go/internal/domain/security"
 	"github.com/gosharplite/tell-me-go/internal/pkg/clock"
+	"github.com/gosharplite/tell-me-go/internal/pkg/metricsfmt"
 )
 
 // sessionManager manages the session lifecycle and agent execution.
@@ -296,14 +297,14 @@ func (o *sessionManager) renderPostTUISummary(ts events.TurnStatus, sd ports.Cha
 			ts.MaxHistoryTokens,
 			ts.Mode, ts.Model)
 	}
-	if metricsLine := ports.FormatMetricsLine(ts); metricsLine != "" {
+	if metricsLine := metricsfmt.FormatMetricsLine(ts); metricsLine != "" {
 		_, _ = fmt.Fprintf(o.Stdout, "%s\n", metricsLine)
 	}
 	turnCost := 0.0
 	if ts.Metrics != nil {
 		turnCost = ts.Metrics.Cost
 	}
-	_, _ = fmt.Fprintf(o.Stdout, "%s\n", ports.FormatFinalLine(ts, turnCost))
+	_, _ = fmt.Fprintf(o.Stdout, "%s\n", metricsfmt.FormatFinalLine(ts, turnCost))
 }
 
 func (o *sessionManager) applyConfiguration(ctx context.Context, chatAgent ports.Chatter, sCfg SessionConfig, sd ports.ChatterComposer, capturer ports.Capturer, tuiOutput bool) (*ui.Bridge, error) {
