+++ internal/pkg/metricsfmt/metrics_format.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2026 gosharplite@gmail.com
 // SPDX-License-Identifier: MIT
 
-package ports
+package metricsfmt
 
 import (
 	"fmt"
